fix: prevent rerender of ListHeaderComponent
Instead of creating an anonymous function for the ListHeaderComponent prop
(which is a new component on every render, causing it to get unmounted and remounted on every render)
render the children as an Element.

According to the documentation of RN FlatList (https://reactnative.dev/docs/flatlist#listheadercomponent),
the ListHeaderComponent can be a function OR an Element. Changing this to element stops the rerenders.